+++ packages/bluesky/lib/src/entities/embed_record.g.dart
@@ -13,7 +13,8 @@ _$_EmbedRecord _$$_EmbedRecordFromJson(Map json) => $checkedCreate(
       json,
       ($checkedConvert) {
         final val = _$_EmbedRecord(
-          type: $checkedConvert(r'$type', (v) => v as String),
+          type: $checkedConvert(
+              r'$type', (v) => v as String? ?? 'app.bsky.embed.record'),
           ref: $checkedConvert('record',
               (v) => StrongRef.fromJson(Map<String, Object?>.from(v as Map))),
         );