+++ packages/bluesky/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added `reply` parameter in `createPost`. ([#246](https://github.com/myConsciousness/atproto.dart/pull/246))
 - Added `facets` parameter in `createPost`. ([#247](https://github.com/myConsciousness/atproto.dart/pull/247))
 - Added `embed` parameter in `createPost`. ([#248](https://github.com/myConsciousness/atproto.dart/pull/248))
+- Supported upload media. ([#249](https://github.com/myConsciousness/atproto.dart/pull/249))
 
 ## v0.3.5